Improved tmux key
The tmux key is one of the keys I use the most and although having it on the thumb improves when used with
the arrow keys it's still uncomfortable, especially when creating a panel or moving to the next one.
For that reason I have chosen to put it next to the main modifiers as a momentary layer toggle with superpowers.
In the tmux layer all key taps will be preceded by ctrl-b except for the arrow keys, which will hold ctrl-b
while the layer is active or one of the other keys is used.